requirements.txt
tcod>=15.0
numpy>=1.24
blinker>=1.6
openai>=0.27
python-dotenv>=1.0
